ATMega328PB support
the ATMega328PB has an extra port, E. This PR adds support for this
port.

I have tested it with a Polulu A-Star 328PB with this board definition
file: https://github.com/pololu/a-star/blob/master/variants/a-star328pb/pins_arduino.h

Thanks for this library, it has saved a lot of legwork and works really
well.