Working with ATtinyx61
Had to merge PCINT0_vect and PCINT1_vect into the single available
interrupt callback, PCINT_vect. Added an initialization method
to clear out interrupts in both PCMSK0 and PCMSK1 on ATtinyx61 as
interrupts on all pins are enabled by default. The split between
PCIE0 and PCIE1 isn't even so a manual range check is done so that
PCIE0 is only used on port 1 for the first 4 pins.